edit_all_set_up

+++ Scraper.jl/src/CSV_edit.jl
@@ -3,6 +3,7 @@ using CSV,DataFrames
 using utils
 using write_utils
 using edit_utils
+using BetterInputFiles
 
 
 function edit_set_up(headers)
@@ -14,17 +15,37 @@ function test_replace()
     replace_known_beginning(content,"Mr Bandt")
 end
 
-function edit_row(row,run_params,header_to_num,row_no)
+function edit_row(row,run_params,header_to_num,row_no,options)
     name_pos = header_to_num["name"]
     row = edit_out_time_content_row(row,header_to_num)
-    if row[name_pos] == "N/A"
-        row[name_pos:name_pos+3] = run_params.previous_talker
+    if options["ADOPT_PREVIOUS_SPEAKER"] 
+        if row[name_pos] == "N/A"
+            row[name_pos:name_pos+3] = run_params.previous_talker
+        end
     end
-    row = remove_the_speaker(row,header_to_num)
-    row = delete_comma(row,header_to_num)
-    row = edit_interjections(row,header_to_num)
-    row = delete_talker_from_content(row,header_to_num)  
+    if options["REMOVE_THE_SPEAKER"] 
+        row = remove_the_speaker(row,header_to_num)
+    end
+    if options["DELETE_SEMICOLON"]
+        row = delete_semicolon(row,header_to_num)
+    end
+    if options["EDIT_INTERJECTIONS"]
+        row = edit_interjections(row,header_to_num)
+    end
+    if options["DELETE_TALKER_FROM_CONTENT"]
+        row = delete_talker_from_content(row,header_to_num) 
+    end
     return row
+end
+
+function set_up_edit_options(fn)
+    input = setup_input("edit_set_up.toml", true)
+    if occursin("question",fn)
+        options = input["QUESTION_TIME_OPTION"]
+    elseif occursin("speech",fn)
+        options = input["SPEECH_OPTION"]
+    end
+    return options
 end
 
 
@@ -35,6 +56,8 @@ function process_csv(fn)
     row_no = 0
     run_params = Edit_Params([" "])
     open("$(fn)_edit.csv", "w") do io
+        options = set_up_edit_options(fn)
+        @show headers
         write_row_to_io(io,headers)
         for row in eachrow(csvfile)
             row_ = @. collect(row)
@@ -44,7 +67,7 @@ function process_csv(fn)
             if talker != "N/A"
                 run_params.previous_talker = row[name_pos:name_pos+3]
             end
-            row = edit_row(row,run_params,header_to_num,row_no)
+            row = edit_row(row,run_params,header_to_num,row_no,options)
             write_row_to_io(io,row)
             row_no += 1
         end
@@ -52,7 +75,11 @@ function process_csv(fn)
 end
 
 function edit_()
-    for fn in ["question_to_answers_1.csv","speeches.csv"]
-        process_csv(fn)
+    for whichchamber in ["fedchamb","chamber"]
+        for fn in ["csvs/$(whichchamber)question_to_answers_1.csv","csvs/$(whichchamber)speeches.csv"]
+            if fn != "csvs/fedchambquestion_to_answers_1.csv"
+                process_csv(fn)
+            end
+        end
     end
 end
